requirements.txt
pandas >1, <2
requests
apache-airflow
pendulum
SQLAlchemy >2
psycopg2-binary
#psycopg[binary]